Hoi Ben je mag dit fixen groetjes Renzo & Hidde :100:

+++ ConsoleApp/SRC/Order.cs
@@ -12,11 +12,11 @@ namespace OptimaliserenPracticum
 		// Variables (these are the values given as input)
 		public int orderNumber, frequency, containerCount, volumePerContainer, matrixID, xCoordinate, yCoordinate;
 		public string placeName;
-		public float emptyingTime;
+		public int emptyingTime;
 		public bool ordersDone = false;
 
 		// Constructor
-		public Order(int orderNumber, string placeName, int frequency, int containerCount, int volumePerContainer, float emptyingTime, int matrixID, int xCoordinate, int yCoordinate)
+		public Order(int orderNumber, string placeName, int frequency, int containerCount, int volumePerContainer, int emptyingTime, int matrixID, int xCoordinate, int yCoordinate)
 		{
 			this.orderNumber = orderNumber;
 			this.placeName = placeName;
